Utilise GITHUB_TOKEN to commit to gh-pages and deploy
Currently, users need to specify a personal access token or a
deploy key to automate the generation of the RepoSense report.
This is an additional step that makes using this repository less
user-friendly for new users.

Let's make use of the autogenerated GITHUB_TOKEN when using
GitHub Actions to automatically push to the gh-pages branch and to
trigger a GitHub Pages build.